Refactor: Split app.py into modular structrure
