+++ app/components/__init__.py
@@ -1,0 +1,15 @@
+"""
+Components package for reusable UI elements
+"""
+from .api_key_manager import ensure_api_key, render_api_key_expander
+from .storage_manager import render_storage_expander
+from .project_settings import render_settings_expander
+from .sidebar import render_sidebar
+
+__all__ = [
+    'ensure_api_key',
+    'render_api_key_expander',
+    'render_storage_expander', 
+    'render_settings_expander',
+    'render_sidebar'
+]
