+++ app/pages/__init__.py
@@ -1,0 +1,12 @@
+"""
+Pages package for main application views
+"""
+from .upload import render_upload_page
+from .dashboard import render_dashboard_page
+from .memo_builder import render_memo_builder_page
+
+__all__ = [
+    'render_upload_page',
+    'render_dashboard_page',
+    'render_memo_builder_page'
+]
